Add PGS tests

+++ lib/src/vm/portable_game_state.dart
@@ -472,7 +472,7 @@ class PortableGameState {
     byteBuffer.add(pvFlag);
 
     // returnvar
-    byteBuffer.add((byte)(discardResult ? 0 : stackFrame.returnVariable));
+    byteBuffer.add(byte(discardResult ? 0 : stackFrame.returnVariable.toInt()));
 
     // argspec
     byteBuffer.add(createArgSpecByte(stackFrame.args));
